requirements.txt
neispy
discord.py